Fixed issue when creating a new job and not specifying presentation links. Fixed issue with matched criteria storage that stopped matches from being generated.

+++ src/main/java/com/avalanche/tmcs/job_posting/JobPostingController.java
@@ -115,18 +115,20 @@ public class JobPostingController {
 
         // Remove existing removed presentation links
         for (JobPresentationLink link : jobPosting.getPresentationLinks()) {
-            if (!link.isInSet(updatedJobPosting.getPresentationLinks())) {
+            if (updatedJobPosting.getPresentationLinks() != null && !link.isInSet(updatedJobPosting.getPresentationLinks())) {
                 jobPosting.getPresentationLinks().remove(link);
                 presentationLinkDAO.delete(link);
             }
         }
 
         // Add new presentation links
-        for (JobPresentationLink link : updatedJobPosting.getPresentationLinks()) {
-            if (!link.isInSet(jobPosting.getPresentationLinks())) {
-                link.setJob(jobPosting);
-                jobPosting.getPresentationLinks().add(link);
-            }
+        if (updatedJobPosting.getPresentationLinks() != null) {
+	        for (JobPresentationLink link : updatedJobPosting.getPresentationLinks()) {
+	            if (!link.isInSet(jobPosting.getPresentationLinks())) {
+	                link.setJob(jobPosting);
+	                jobPosting.getPresentationLinks().add(link);
+	            }
+	        }
         }
 
         jobPostingDAO.save(jobPosting);
